Add '$' to the end of some patterns
When looking for the shell prompt, adding '$' to the end of some
patterns should make the matching more precise.  If there are two
prompts in the expect buffer, then only the last one will count as the
match, which is what we want in all of the cases I have added the '$'.