run-test.pl: Clear exec state hash before restarting a process.
The exec-state hashes are used to track the state of an executing
sub-process. When the sub-process dies the hash is updated in the
SIGCHLD handler.

When we later restart the process, we would previusly fail to clear
out the exec-state hash, with the result that when the child exited
(or died) a second time the hash would already contain state
indicating that the process had died, this would trigger an assert.

The correct solution is that when we know the child process is about
to be restarted, clear out the exec-state hash, ensuring that each
time the child process exits it will look like the first time.