run-tests.pl: exec_runtest_process improvements
Don't pass the IP address in as a separate parameter, this can be
fetched from the simulator object.

Fill in the header comment on exec_runtest_process.

Redirect the stdout and stderr of exec_runtest_process to files within
the results directory.  This means the output does not clutter up the
output trace of the main run-tests.pl script, and the output is
recorded for reference.  For now I've put both stdout and stderr into
a single file (runtest.output) for each exec of runtest.  This should
be fine as runtest is pretty quiet on stderr unless there's a problem.