run-tests.pl: improve function exec_simulator_process
Fill in the header comment for functino exec_simulator_process.

Block SIGINT and SIGCHLD around the fork within the function, to avoid
signals arriving while data structures are in flux.

Redirect stdout and stderr of the simulator into a file to avoid
cluttering the output of the top level runtests script.  Also this
keeps the output around if it is ever needed.  During standard
operation the simulator does not seem to produce any output to either
stdout, or stderr.  I guess this might change if there's any problems.